+++ src/schema/interfaces.tsx
@@ -44,7 +44,7 @@ export interface ClassificationProps extends SpaProps {
 }
 
 export interface BaseProcessingJobProps {
-  autoProc: AutoProcSchema;
+  autoProc: AutoProcSchema | null;
   procJob: ProcessingJobSchema;
   status: string;
   active: boolean;